+++ src/main/java/br/com/infnet/services/OrganizadorServiceImpl.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import br.com.infnet.exceptions.NegocioException;
+import br.com.infnet.exceptions.TabelaDeErros;
 import br.com.infnet.models.Organizador;
 import br.com.infnet.repositorys.OrganizadorRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class OrganizadorServiceImpl implements OrganizadorManager {
+
 	private final OrganizadorRepository organizadorRepository;
 
 	@Override
@@ -20,21 +23,31 @@ public class OrganizadorServiceImpl implements OrganizadorManager {
 
 	@Override
 	public void atualizarOrganizador(Organizador organizador) {
+		if (!organizadorRepository.existsById(organizador.getId())) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
 		organizadorRepository.save(organizador);
 	}
 
 	@Override
 	public void deletarOrganizador(Long id) {
+		if (!organizadorRepository.existsById(id)) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
 		organizadorRepository.deleteById(id);
 	}
 
 	@Override
 	public Organizador buscarOrganizadorPorId(Long id) {
-		return organizadorRepository.findById(id).orElse(null);
+		return organizadorRepository.findById(id).orElseThrow(() -> new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO));
 	}
 
 	@Override
 	public List<Organizador> listarOrganizadores() {
-		return organizadorRepository.findAll();
+		List<Organizador> organizadores = organizadorRepository.findAll();
+		if (organizadores.isEmpty()) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
+		return organizadores;
 	}
 }
