+++ src/main/java/br/com/infnet/services/ParticipanteServiceImpl.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import br.com.infnet.exceptions.NegocioException;
+import br.com.infnet.exceptions.TabelaDeErros;
 import br.com.infnet.models.Participante;
 import br.com.infnet.repositorys.ParticipanteRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +13,7 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class ParticipanteServiceImpl implements ParticipanteManager {
+
 	private final ParticipanteRepository participanteRepository;
 
 	@Override
@@ -20,21 +23,31 @@ public class ParticipanteServiceImpl implements ParticipanteManager {
 
 	@Override
 	public void atualizarParticipante(Participante participante) {
+		if (!participanteRepository.existsById(participante.getId())) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
 		participanteRepository.save(participante);
 	}
 
 	@Override
 	public void deletarParticipante(Long id) {
+		if (!participanteRepository.existsById(id)) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
 		participanteRepository.deleteById(id);
 	}
 
 	@Override
 	public Participante buscarParticipantePorId(Long id) {
-		return participanteRepository.findById(id).orElse(null);
+		return participanteRepository.findById(id).orElseThrow(() -> new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO));
 	}
 
 	@Override
 	public List<Participante> listarParticipantes() {
-		return participanteRepository.findAll();
+		List<Participante> participantes = participanteRepository.findAll();
+		if (participantes.isEmpty()) {
+			throw new NegocioException(TabelaDeErros.REGISTRO_NAO_ENCONTRADO);
+		}
+		return participantes;
 	}
 }
